Fold recolor-sst into sst-overlay.
